requirements.txt
APScheduler==3.0.1
Flask==0.10.1
Flask-API==0.6.2
Flask-Compress==1.0.2
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
Werkzeug==0.9.6
cffi==0.8.6
cryptography==0.6
futures==2.2.0
gunicorn==19.1.1
itsdangerous==0.24
pyOpenSSL==0.14
pycparser==2.10
pymongo==2.7.2
pytz==2014.10
requests==2.4.3
six==1.9.0
tzlocal==1.1.2
wsgiref==0.1.2
